+++ src/Mpociot/BotMan/Interfaces/QuestionActionInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Mpociot\BotMan\Interfaces;
+
+interface QuestionActionInterface
+{
+    /**
+     * Array representation of the question action.
+     *
+     * @return array
+     */
+    public function toArray();
+}